Remove Local Server Call on Logistration LEARNER-966
